+++ src/main/java/com/shinhan/memento/controller/MentosController.java
@@ -1,28 +1,11 @@
 package com.shinhan.memento.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import com.shinhan.memento.common.exception.MentosException;
-import com.shinhan.memento.common.response.BaseResponse;
-import com.shinhan.memento.common.response.status.BaseExceptionResponseStatus;
-import com.shinhan.memento.dto.CreateMentosDTO;
-import com.shinhan.memento.dto.MentosDTO;
-import com.shinhan.memento.service.MentosService;
-
-import lombok.extern.slf4j.Slf4j;
-
-@RestController
+@Controller
 @RequestMapping("/mentos")
-@Slf4j
 public class MentosController {
-
-	@Autowired
-	private MentosService mentosService;
 
 	@RequestMapping("/full")
 	public String mentosAllView() {
@@ -58,17 +41,5 @@ public class MentosController {
 	public String mentosMentoDetailClassList() {
 		return "mentos/mentosMentoDetailClassList";
 	}
-
-	@PostMapping("/creatementos")
-	public BaseResponse<Void> createMento(@RequestBody CreateMentosDTO dto) {
-		log.info("MentosController.createMento");
-		boolean result = mentosService.registerMento(dto);
-		if (!result) {
-			throw new MentosException(BaseExceptionResponseStatus.CANNOT_CREATE_MENTOS);
-		}
-		return new BaseResponse<>(null);
-	}
-	
-
 
 }
